+++ ISH-CorrespondenceService/src/main/java/in/lakshay/entity/ElgibilityDetailsEntity.java
@@ -17,7 +17,7 @@ public class ElgibilityDetailsEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer edTraceId;
-    private Integer caseNo;
+    private Long caseNo;
     @Column(length = 30)
     private String holderName;
     private Long holderSSN;
