+++ ISH-CorrespondenceService/src/main/java/in/lakshay/repository/ICOTriggerRepository.java
@@ -9,5 +9,5 @@ import in.lakshay.entity.CoTriggersEntity;
 public interface ICOTriggerRepository extends JpaRepository<CoTriggersEntity, Integer> {
     public List<CoTriggersEntity> findByTriggerStatus(String status);
 
-    public CoTriggersEntity findByCaseNo(int caseNo);
+    public CoTriggersEntity findByCaseNo(Long caseNo);
 }
