+++ ISH-CorrespondenceService/src/main/java/in/lakshay/repository/IElgibiltyDeterminationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import in.lakshay.entity.ElgibilityDetailsEntity;
 
 public interface IElgibiltyDeterminationRepository extends JpaRepository<ElgibilityDetailsEntity, Integer> {
-    public ElgibilityDetailsEntity findByCaseNo(int caseno);
+    public ElgibilityDetailsEntity findByCaseNo(Long caseno);
 }
