+++ ISH-DataCollectionService/src/main/java/in/lakshay/bindings/EducationInputs.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class EducationInputs {
 	private Integer educationId;
-	private Integer caseNo;
+	private Long caseNo;
 	private String highestQlfy;
 	private Integer passOutYear;
