+++ ISH-DataCollectionService/src/main/java/in/lakshay/bindings/PlanSelectionInputs.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PlanSelectionInputs {
 	private Integer planId;
-	private Integer caseNo;
+	private Long caseNo;
 	private Integer appId;
 	private String planName;
