+++ ISH-DataCollectionService/src/main/java/in/lakshay/entity/DcChildrenEntity.java
@@ -13,11 +13,11 @@ import lombok.Data;
 @Table(name="DC_CHILDREN")
 @Data
 public class DcChildrenEntity {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Integer childId;
-	private Integer caseNo;
+	private Long caseNo;
 	private LocalDate childDOB;
 	private Long childSSN;
